apps/mobthresh/requirements.txt
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
